test (global) : 분산락 컴포넌트 테스트코드 수정
